Unibz Master's Thesis Project: Unimatch
This is a UNIBZ research project devoted to personalized rankings and recommendations in the educational domain. Respondent is asked to rate a few universities that are offered based on the geographic region they select.  Based on these ratings the system would show three lists of potential universities the respondent might be interested in.

This Master thesis is undertaken by Anna Lambrix with the supervision of the project by Nabil El Ioini and Mehdi Elahi.